+++ R/oauth-refresh.R
@@ -27,12 +27,18 @@ refresh_oauth2.0 <- function(endpoint, app, credentials, user_params = NULL,
     response <- POST(refresh_url, body = req_params, encode = "form")
   }
 
-  if (known_oauth2.0_error(response)) {
-    warning("Unable to refresh token", call. = FALSE)
+  err <- find_oauth2.0_error(response)
+  if (!is.null(err)) {
+    lines <- c(
+      paste0("Unable to refresh token: ", err$error),
+      err$error_description,
+      err$error_uri
+    )
+    warning(paste(lines, collapse = "\n"), call. = FALSE)
     return(NULL)
   }
-  stop_for_status(response)
 
+  stop_for_status(response)
   refresh_data <- content(response)
   utils::modifyList(credentials, refresh_data)
 }